# userservices/.optic/generated/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Optic Generated OpenAPI",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/patient/getInvoiceById": {
      "get": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "usr": {
                    "type": "string"
                  }
                },
                "required": [
                  "usr"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "uploadPAN": {
                            "type": "string"
                          },
                          "uploadIncomeProof": {
                            "type": "string"
                          },
                          "uploadHospitalBill": {
                            "type": "string"
                          },
                          "uploadCancelledCheque": {
                            "type": "string"
                          },
                          "uploadBankStatement": {
                            "type": "string"
                          },
                          "uploadAadharFront": {
                            "type": "string"
                          },
                          "uploadAadharBack": {
                            "type": "string"
                          },
                          "updatedBy": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          },
                          "totalIncome": {
                            "type": "number"
                          },
                          "scheme": {
                            "type": "number"
                          },
                          "repaymentData": {
                            "type": "array",
                            "items": {}
                          },
                          "relationship": {
                            "type": "string"
                          },
                          "relation": {
                            "nullable": true
                          },
                          "referenceName": {
                            "type": "string"
                          },
                          "processingFees": {
                            "type": "number"
                          },
                          "permanentState": {
                            "type": "string"
                          },
                          "permanentPinCode": {
                            "type": "number"
                          },
                          "permanentCity": {
                            "type": "string"
                          },
                          "permanentAddress": {
                            "type": "string"
                          },
                          "patientName": {
                            "type": "string"
                          },
                          "organizationName": {
                            "type": "string"
                          },
                          "organizationId": {
                            "type": "string"
                          },
                          "occupation": {
                            "type": "string"
                          },
                          "loanAmount": {
                            "type": "number"
                          },
                          "isInsured": {
                            "type": "boolean"
                          },
                          "isDeleted": {
                            "type": "boolean"
                          },
                          "isActive": {
                            "type": "boolean"
                          },
                          "invoiceStatus": {
                            "type": "string"
                          },
                          "invoiceId": {
                            "type": "number"
                          },
                          "interest_Amount": {
                            "type": "number"
                          },
                          "interest": {
                            "type": "number"
                          },
                          "hospitalName": {
                            "type": "string"
                          },
                          "emailIdRefPerson": {
                            "type": "string"
                          },
                          "emailId": {
                            "type": "string"
                          },
                          "deduction_Amount": {
                            "type": "number"
                          },
                          "dateOfBirth": {
                            "type": "string"
                          },
                          "currentState": {
                            "type": "string"
                          },
                          "currentPinCode": {
                            "type": "number"
                          },
                          "currentCity": {
                            "type": "string"
                          },
                          "currentAddress": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "contactNumber": {
                            "type": "number"
                          },
                          "contactNoRefPerson": {
                            "type": "number"
                          },
                          "companyName": {
                            "type": "string"
                          },
                          "branch": {
                            "type": "string"
                          },
                          "bankAssociated": {
                            "type": "string"
                          },
                          "accountNumber": {
                            "type": "number"
                          },
                          "_id": {
                            "type": "string"
                          },
                          "__v": {
                            "type": "number"
                          },
                          "PF_Amount": {
                            "type": "number"
                          },
                          "IFSCCode": {
                            "type": "string"
                          },
                          "GST_Amount": {
                            "type": "number"
                          },
                          "EMI_Tenure": {
                            "type": "number"
                          },
                          "EMI_Amount": {
                            "type": "number"
                          },
                          "Disbursed_Amount": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "uploadPAN",
                          "uploadIncomeProof",
                          "uploadHospitalBill",
                          "uploadCancelledCheque",
                          "uploadBankStatement",
                          "uploadAadharFront",
                          "uploadAadharBack",
                          "updatedBy",
                          "updatedAt",
                          "totalIncome",
                          "scheme",
                          "repaymentData",
                          "relationship",
                          "relation",
                          "referenceName",
                          "processingFees",
                          "permanentState",
                          "permanentPinCode",
                          "permanentCity",
                          "permanentAddress",
                          "patientName",
                          "organizationName",
                          "organizationId",
                          "occupation",
                          "loanAmount",
                          "isInsured",
                          "isDeleted",
                          "isActive",
                          "invoiceStatus",
                          "invoiceId",
                          "interest_Amount",
                          "interest",
                          "hospitalName",
                          "emailIdRefPerson",
                          "emailId",
                          "deduction_Amount",
                          "dateOfBirth",
                          "currentState",
                          "currentPinCode",
                          "currentCity",
                          "currentAddress",
                          "createdAt",
                          "contactNumber",
                          "contactNoRefPerson",
                          "companyName",
                          "branch",
                          "bankAssociated",
                          "accountNumber",
                          "_id",
                          "__v",
                          "PF_Amount",
                          "IFSCCode",
                          "GST_Amount",
                          "EMI_Tenure",
                          "EMI_Amount",
                          "Disbursed_Amount"
                        ]
                      }
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "errors"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}